+++ src/app/help/faq/components/TitledSection.tsx
